+++ source/Messaging.Application/Transactions/MoveIn/IMoveInTransactionRepository.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Messaging.Application.Transactions
+namespace Messaging.Application.Transactions.MoveIn
 {
     /// <summary>
     /// Storage for transactions
