+++ source/Messaging.Application/Transactions/MoveIn/MoveInTransaction.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Messaging.Application.Transactions
+namespace Messaging.Application.Transactions.MoveIn
 {
     public class MoveInTransaction
     {
