+++ source/Messaging.Infrastructure/Transactions/MoveInTransactionRepository.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Messaging.Application.Transactions;
+using Messaging.Application.Transactions.MoveIn;
 using Messaging.Infrastructure.Configuration.DataAccess;
 
 namespace Messaging.Infrastructure.Transactions
