+++ source/Messaging.IntegrationTests/IncomingMessages/RequestChangeOfSupplierHandlingTests.cs
@@ -19,6 +19,7 @@ using Messaging.Application.Common;
 using Messaging.Application.IncomingMessages.RequestChangeOfSupplier;
 using Messaging.Application.OutgoingMessages;
 using Messaging.Application.Transactions;
+using Messaging.Application.Transactions.MoveIn;
 using Messaging.Application.Xml;
 using Messaging.Application.Xml.SchemaStore;
 using Messaging.IntegrationTests.Fixtures;
